pyproject.toml
[project]
name = "clinic-copilot-nlp"
version = "0.1.0"
description = "FastAPI backend for clinical history summarization using LLMs"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "openai>=1.54.0",
    "pydantic-settings>=2.6.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.4.10",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
